requirements.txt
crewai>=1.8.1
python-dotenv>=1.0.0
